#22 [In Progress] add user session and role check before performing POST, PUT and DELETE requests on clinic objects

+++ dentalHeathcareSystem/dentist_clinic_service/Controller/clinicController.js
@@ -5,25 +5,46 @@ const MqttBroker = require("../mqtt-broker");
 
 
 exports.clinicCreate = async (payload) => {
+    // console.log('incoming payload ', payload.toString());
+    // parsedPayload contains both object attributes and session variables. try the to print the incoming payload to see the session variables
+    /* example payload: clinic payload  {
+    "name":"Tandkliniken Johanneberg",
+    "address":"Universitetsplatsen 1, 405 30 Göteborg",
+    "email":"johanneberg@gmail.com",
+    "phoneNumber":"1234567890",
+    "userId":"67630274c5502c7c3cb1b318",
+    "role":"admin"} */
     try {
         var message ='';
         var status = 0;
-        const newClinic = JSON.parse(payload);
-    
-        const newClinicValidation = validateClinic(newClinic);
-        if(!newClinicValidation.success) {
-            status = 400
-            return status +"/"+ newClinicValidation.message;
-        }
-     
-        const clinic = new Clinic(newClinic);
-        await clinic.save();
-        message = "Clinic created"
-        var stringClinic = JSON.stringify(clinic) 
-        status = 200;
-        return status +"/"+ message +"/"+ stringClinic;
+        const parsedPayload = JSON.parse(payload);
+        const userRole = parsedPayload.role;
 
-    } catch(error) {
+        /* only admin can create a clinic */
+        if (userRole === 'admin') {
+            const newClinicValidation = validateClinic(parsedPayload);
+            if(!newClinicValidation.success) {
+                status = 400
+                return status +"/"+ newClinicValidation.message;
+            }
+        
+            const clinic = new Clinic(parsedPayload);
+            await clinic.save();
+
+            var stringClinic = JSON.stringify(clinic);
+            message = "Clinic created"
+            status = 200;
+            return status + "/" + message + "/" + stringClinic;
+            
+        } else {
+            message = "Unauthorized request. Only admin can perform this action. "
+            status = 400;
+            return status +"/"+ message;
+        }
+        
+
+
+    } catch (error) {        
         status = 400
         return status +"/"+ error.message;
     }
@@ -79,35 +100,44 @@ exports.updateAClinic = async (topic,payload) => {
         var message ='';
         var status = 0;
         var topicArr = topic.split("/");
-        const id = topicArr[2];; 
+        const id = topicArr[2];
 
-        var clinic = JSON.parse(payload);
+        var parsedPayload = JSON.parse(payload);
+        const userRole = parsedPayload.role;
 
-        const existingClinic = await Clinic.findById(id);
+        /* only admin can update a clinic */
+        if (userRole === 'admin') { 
+            const existingClinic = await Clinic.findById(id);
 
 
-        var name = clinic.name ? clinic.name : existingClinic.name;
-        var email = clinic.email ? clinic.email : existingClinic.email;
-        var phoneNumber = clinic.phoneNumber ? clinic.phoneNumber : existingClinic.phoneNumber;        
-        var address = clinic.address ? clinic.address : existingClinic.address;
-        var dentists = clinic.dentists ? clinic.dentists : existingClinic.dentists;
-        var appointments = clinic.appointments ? clinic.appointments : existingClinic.appointments;
-        
-        var updatedClinic = await Clinic.findByIdAndUpdate(id, {
-            name: name,
-            email: email,
-            phoneNumber: phoneNumber,
-            address: address, 
-            dentists: dentists,
-            appointments: appointments
-        },
-            {new: true}
-        );
-        await updatedClinic.save()
-            status = 200;
-            message = "Clinic information updated successfully";
-            var stringClinic = JSON.stringify(updatedClinic);
-            return status +"/"+ message +"/"+ stringClinic;
+            var name = parsedPayload.name ? parsedPayload.name : existingClinic.name;
+            var email = parsedPayload.email ? parsedPayload.email : existingClinic.email;
+            var phoneNumber = parsedPayload.phoneNumber ? parsedPayload.phoneNumber : existingClinic.phoneNumber;        
+            var address = parsedPayload.address ? parsedPayload.address : existingClinic.address;
+            var dentists = parsedPayload.dentists ? parsedPayload.dentists : existingClinic.dentists;
+            var appointments = parsedPayload.appointments ? parsedPayload.appointments : existingClinic.appointments;
+            
+            var updatedClinic = await Clinic.findByIdAndUpdate(id, {
+                name: name,
+                email: email,
+                phoneNumber: phoneNumber,
+                address: address, 
+                dentists: dentists,
+                appointments: appointments
+            },
+                {new: true}
+            );
+            await updatedClinic.save()
+                status = 200;
+                message = "Clinic information updated successfully";
+                var stringClinic = JSON.stringify(updatedClinic);
+                return status +"/"+ message +"/"+ stringClinic;
+        }else {
+            message = "Unauthorized request. Only admin can perform this action. "
+            status = 400;
+            return status +"/"+ message;
+        }
+
 
     } catch (error) {
         status = 400; 
@@ -146,22 +176,34 @@ exports.getDentistFromClinic = async (topic) => {
   }
 };
 
-exports.deleteAClinic = async (topic) => {
-    try{
-        var message ='';
-        var status = 0;
-        var topicArr = topic.split("/");
-        const id = topicArr[2];
-        const clinic = await Clinic.findByIdAndDelete(id);
-        if(!clinic){
-            status = 404
-            message = "No clinic found";
-            return status +"/"+ message; 
+exports.deleteAClinic = async (topic, payload) => {
+    try {
+        var parsedPayload = JSON.parse(payload);
+        const userRole = parsedPayload.role;
+
+        /* only admin can create a clinic */
+        if (userRole === 'admin') { 
+            var message ='';
+            var status = 0;
+            var topicArr = topic.split("/");
+            const id = topicArr[2];
+            const clinic = await Clinic.findByIdAndDelete(id);
+            if(!clinic){
+                status = 404
+                message = "No clinic found";
+                return status +"/"+ message; 
+            }
+            status =200;
+            message = "Clinic deleted";
+            var stringClinic = JSON.stringify(clinic);
+            return status + "/" + message + "/" + stringClinic;
+            
+        }else {
+            message = "Unauthorized request. Only admin can perform this action. "
+            status = 400;
+            return status +"/"+ message;
         }
-        status =200;
-        message = "Clinic deleted";
-        var stringClinic = JSON.stringify(clinic);
-        return status +"/"+ message +"/"+ stringClinic;
+
 
     }catch (error) {
         status = 400;
